+++ torch/_inductor/dtype_propagation.py
@@ -19,7 +19,7 @@ if TYPE_CHECKING:
 
 import torch
 from torch._inductor.virtualized import V
-from torch._prims_common import ELEMENTWISE_TYPE_PROMOTION_KIND
+from torch._prims_common import ELEMENTWISE_TYPE_PROMOTION_KIND, type_to_dtype
 
 from .utils import upcast_compute_type
 from .virtualized import OpsValue
@@ -72,7 +72,8 @@ def promote_types(
 
     for arg in args:
         if isinstance(arg, str):
-            # comes from templates.. TODO
+            # TODO: fix the flex attention instances
+            assert isinstance(V.get_ops_handler(), torch._inductor.select_algorithm.ModificationWrapper)
             continue
 
         if isinstance(arg, OpsValue):
@@ -80,7 +81,7 @@ def promote_types(
             assert isinstance(arg, torch._prims_common.Number) or hasattr(arg, "dtype")
 
         if isinstance(arg, torch._prims_common.Number):
-            dtype_prop_candidates.append((torch.tensor(arg).dtype, True))
+            dtype_prop_candidates.append((type_to_dtype(type(arg)), True))
             continue
 
         dtype_prop_candidates.append((arg.dtype, False))
